+++ src/main/java/TeamDPlus/code/service/artwork/like/ArtworkLikeService.java
@@ -1,6 +1,7 @@
 package TeamDPlus.code.service.artwork.like;
 
 
+import TeamDPlus.code.advice.ApiRequestException;
 import TeamDPlus.code.domain.account.Account;
 import TeamDPlus.code.domain.artwork.ArtWorkRepository;
 import TeamDPlus.code.domain.artwork.ArtWorks;
@@ -19,9 +20,9 @@ public class ArtworkLikeService {
 
     @Transactional
     public void doLike(Account account, Long artWorkId) {
-        ArtWorks artWorks = artWorkRepository.findById(artWorkId).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시글 입니다."));
+        ArtWorks artWorks = getArtWorks(artWorkId);
         if (artWorkLikesRepository.existByAccountIdAndArtWorkId(account.getId(), artWorkId)) {
-            throw new IllegalArgumentException("이미 좋아요한 게시글 입니다.");
+            throw new ApiRequestException("이미 좋아요한 게시글 입니다.");
         }
         artWorks.getAccount().getRank().upRankScore();
         ArtWorkLikes artWorkLikes = ArtWorkLikes.builder().artWorks(artWorks).account(account).build();
@@ -30,9 +31,16 @@ public class ArtworkLikeService {
 
     @Transactional
     public void unLike(Account account, Long artWorkId) {
-        ArtWorks artWorks = artWorkRepository.findById(artWorkId).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시글 입니다."));
+        ArtWorks artWorks = getArtWorks(artWorkId);
+        if (!artWorkLikesRepository.existByAccountIdAndArtWorkId(account.getId(), artWorkId)) {
+            throw new ApiRequestException("이미 좋아요 해지한 게시글 입니다.");
+        }
         artWorks.getAccount().getRank().downRankScore();
         artWorkLikesRepository.deleteByArtWorksIdAndAccountId(artWorkId,account.getId());
+    }
+
+    private ArtWorks getArtWorks(Long artWorkId) {
+        return artWorkRepository.findById(artWorkId).orElseThrow(() -> new ApiRequestException("존재하지 않는 게시글 입니다."));
     }
